+++ v3/crates/metadata-resolve/src/stages/data_connectors/mod.rs
@@ -29,12 +29,15 @@ pub fn resolve<'a>(
         let qualified_data_connector_name =
             Qualified::new(subgraph.clone(), data_connector.name.clone());
 
-        let (data_connector_context, connector_issues) =
-            types::DataConnectorContext::new(data_connector, &configuration.unstable_features)
-                .map_err(|error| NamedDataConnectorError {
-                    data_connector_name: qualified_data_connector_name.clone(),
-                    error,
-                })?;
+        let (data_connector_context, connector_issues) = types::DataConnectorContext::new(
+            metadata_accessor,
+            data_connector,
+            &configuration.unstable_features,
+        )
+        .map_err(|error| NamedDataConnectorError {
+            data_connector_name: qualified_data_connector_name.clone(),
+            error,
+        })?;
 
         issues.extend(
             connector_issues