+++ v3/crates/metadata-resolve/src/stages/type_permissions/mod.rs
@@ -12,6 +12,7 @@ use crate::types::subgraph::Qualified;
 
 use crate::helpers::typecheck;
 use crate::stages::object_types;
+use crate::ValueExpressionOrPredicate;
 
 /// resolve type permissions
 pub fn resolve(
@@ -54,6 +55,7 @@ pub fn resolve(
                     output_type_permission,
                 )?;
                 object_type.type_input_permissions = resolve_input_type_permission(
+                    &metadata_accessor.flags,
                     &object_type.object_type,
                     output_type_permission,
                 )?;
@@ -97,6 +99,7 @@ pub fn resolve_output_type_permission(
 }
 
 pub(crate) fn resolve_input_type_permission(
+    flags: &open_dds::flags::Flags,
     object_type_representation: &object_types::ObjectTypeRepresentation,
     type_permissions: &TypePermissionsV1,
 ) -> Result<BTreeMap<Role, TypeInputPermission>, TypeInputPermissionError> {
@@ -133,10 +136,23 @@ pub(crate) fn resolve_input_type_permission(
                         );
                     }
                 };
+                let resolved_value = match &value {
+                    open_dds::permissions::ValueExpression::Literal(literal) => {
+                        ValueExpressionOrPredicate::Literal(literal.clone())
+                    }
+                    open_dds::permissions::ValueExpression::SessionVariable(session_variable) => {
+                        ValueExpressionOrPredicate::SessionVariable(
+                            hasura_authn_core::SessionVariableReference {
+                                name: session_variable.clone(),
+                                passed_as_json: flags.json_session_variables,
+                            },
+                        )
+                    }
+                };
                 resolved_field_presets.insert(
                     field_name.clone(),
                     FieldPresetInfo {
-                        value: value.clone(),
+                        value: resolved_value,
                         deprecated: field_definition.deprecated.clone(),
                     },
                 );