+++ v3/crates/metadata-resolve/src/stages/type_permissions/types.rs
@@ -1,12 +1,12 @@
 use std::collections::BTreeMap;
 
 use open_dds::{
-    permissions::{Role, TypeOutputPermission, ValueExpression},
+    permissions::{Role, TypeOutputPermission},
     types::Deprecated,
 };
 
-use crate::stages::object_types;
 use crate::Qualified;
+use crate::{stages::object_types, ValueExpressionOrPredicate};
 use open_dds::types::{CustomTypeName, FieldName};
 use serde::{Deserialize, Serialize};
 use std::ops::Deref;
@@ -43,7 +43,7 @@ pub struct TypeInputPermission {
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub struct FieldPresetInfo {
-    pub value: ValueExpression,
+    pub value: ValueExpressionOrPredicate,
     pub deprecated: Option<Deprecated>,
 }