+++ v3/crates/metadata-resolve/src/types/permission.rs
@@ -4,12 +4,12 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub enum ValueExpression {
     Literal(serde_json::Value),
-    SessionVariable(open_dds::session_variables::SessionVariable),
+    SessionVariable(open_dds::session_variables::SessionVariableReference),
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub enum ValueExpressionOrPredicate {
     Literal(serde_json::Value),
-    SessionVariable(open_dds::session_variables::SessionVariable),
+    SessionVariable(open_dds::session_variables::SessionVariableReference),
     BooleanExpression(Box<model_permissions::ModelPredicate>),
 }